+++ python/akquant/__init__.py
@@ -8,7 +8,7 @@ try:
 except metadata.PackageNotFoundError:
     __version__ = "0.0.0+unknown"
 
-__engine_rule_version__ = "1.3.0"  # Increment on behavior-changing updates
+__engine_rule_version__ = "1.3.1"  # Increment on behavior-changing updates
 
 from . import akquant as _akquant
 from . import log as _log